requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Authlib==1.3.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
cryptography==41.0.7
python-dotenv==1.0.0
python-dateutil==2.8.2
hypothesis==6.92.1
pytest==7.4.3
requests==2.31.0
Flask-CORS==4.0.0
PyJWT==2.8.0
